@@ -2,12 +2,12 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\..\..\csharp\FlutterCsharpRpc.csproj" />
+		<ProjectReference Include="..\..\..\csharp\FlutterSharpRpc.csproj" />
 	</ItemGroup>
 
 </Project>
